LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/text/2d.text.font.parse.basic-expected.txt

2d.text.font.parse.basic


FAIL OffscreenCanvas test: 2d.text.font.parse.basic assert_equals: ctx.font === '20px serif' (got 20PX   SERIF[string], expected 20px serif[string]) expected "20px serif" but got "20PX   SERIF"
